requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
Babel==2.9.1
boto3==1.21.4
botocore==1.24.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
detect-delimiter==0.1.1
device-detector==5.0.1
Django==4.0.2
django-cors-headers==3.11.0
django-csp==3.7
django-hosts==5.1
django-ipware==4.0.2
django-polymorphic==3.1.0
django-recaptcha==3.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
install==1.3.5
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.1.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
urllib3==1.26.8
